feat(laravel): split render logic from error handler
